preserve enums in types
