+++ plugins/cost-insights/src/components/PeriodSelect/PeriodSelect.test.tsx
@@ -20,9 +20,8 @@ import { renderInTestApp } from '@backstage/test-utils';
 import UserEvent from '@testing-library/user-event';
 import { PeriodSelect, getDefaultOptions } from './PeriodSelect';
 import { getDefaultPageFilters } from '../../utils/filters';
-import { Duration } from '../../utils/duration';
 import { MockBillingDateProvider } from '../../utils/tests';
-import { Group } from '../../types';
+import { Group, Duration } from '../../types';
 
 const DefaultPageFilters = getDefaultPageFilters([{ id: 'tools' }] as Group[]);
 const lastCompleteBillingDate = '2020-05-01';
