+++ plugins/cost-insights/src/components/ProductInsightsCard/ProductInsightsCard.test.tsx
@@ -35,8 +35,7 @@ import {
   MockScrollProvider,
   MockLoadingProvider,
 } from '../../utils/tests';
-import { Duration } from '../../utils/duration';
-import { Product, ProductCost, ProductPeriod } from '../../types';
+import { Duration, Product, ProductCost, ProductPeriod } from '../../types';
 
 const costInsightsApi = (
   productCost: ProductCost,
