+++ plugins/cost-insights/src/components/ProductInsightsCard/ProductInsightsCard.tsx
@@ -28,10 +28,9 @@ import {
   useLoading,
   useScroll,
 } from '../../hooks';
-import { Duration } from '../../utils/duration';
 import { useProductInsightsCardStyles as useStyles } from '../../utils/styles';
 import { mapFiltersToProps, mapLoadingToProps } from './selector';
-import { Maybe, Product, ProductCost } from '../../types';
+import { Duration, Maybe, Product, ProductCost } from '../../types';
 import { pluralOf } from '../../utils/grammar';
 import { formatPeriod } from '../../utils/formatters';
