+++ plugins/cost-insights/src/components/ResourceGrowthBarChartLegend/ResourceGrowthBarChartLegend.test.tsx
@@ -17,10 +17,10 @@
 import React, { PropsWithChildren } from 'react';
 import { renderInTestApp } from '@backstage/test-utils';
 import { ResourceGrowthBarChartLegend } from './ResourceGrowthBarChartLegend';
-import { Duration } from '../../utils/duration';
 import { defaultCurrencies } from '../../utils/currency';
 import { findAlways } from '../../utils/assert';
 import { MockConfigProvider, MockCurrencyProvider } from '../../utils/tests';
+import { Duration } from '../../types';
 
 const engineers = findAlways(defaultCurrencies, c => c.kind === null);
