+++ plugins/cost-insights/src/components/ResourceGrowthBarChartLegend/ResourceGrowthBarChartLegend.tsx
@@ -19,8 +19,7 @@ import { Box, useTheme } from '@material-ui/core';
 import { LegendItem } from '../LegendItem';
 import { CostGrowth } from '../CostGrowth';
 import { currencyFormatter } from '../../utils/formatters';
-import { Duration } from '../../utils/duration';
-import { ChangeStatistic, CostInsightsTheme } from '../../types';
+import { ChangeStatistic, Duration, CostInsightsTheme } from '../../types';
 
 export type ResourceGrowthBarChartLegendProps = {
   change: ChangeStatistic;
