+++ plugins/cost-insights/src/index.ts
@@ -19,4 +19,5 @@ export * from './api';
 export * from './components';
 export { useCurrency } from './hooks';
 export * from './types';
-export * from './utils/public';
+export * from './utils/tests';
+export * from './utils/duration';
