+++ plugins/cost-insights/src/types/Alert.ts
@@ -14,11 +14,8 @@
  * limitations under the License.
  */
 
-import React from 'react';
 import { ChangeStatistic } from './ChangeStatistic';
 import { Maybe } from './Maybe';
-import { UnlabeledDataflowAlertCard } from '../components/UnlabeledDataflowAlertCard';
-import { ProjectGrowthAlertCard } from '../components/ProjectGrowthAlertCard';
 
 /**
  * Generic alert type with required fields for display. The `element` field will be rendered in
@@ -57,11 +54,6 @@ export enum DataKey {
   Name = 'name',
 }
 
-/**
- * The alerts below are examples of Alert implementation; the CostInsightsApi permits returning
- * any implementation of the Alert type, so adopters can create their own. The CostInsightsApi
- * fetches alert data from the backend, then creates Alert classes with the data.
- */
 export interface ProjectGrowthData {
   project: string;
   periodStart: string;
@@ -71,49 +63,12 @@ export interface ProjectGrowthData {
   products: Array<AlertCost>;
 }
 
-export class ProjectGrowthAlert implements Alert {
-  data: ProjectGrowthData;
-
-  constructor(data: ProjectGrowthData) {
-    this.data = data;
-  }
-
-  get title() {
-    return `Investigate cost growth in project ${this.data.project}`;
-  }
-
-  subtitle =
-    'Cost growth outpacing business growth is unsustainable long-term.';
-  url = '/cost-insights/investigating-growth';
-
-  get element() {
-    return <ProjectGrowthAlertCard alert={this.data} />;
-  }
-}
-
 export interface UnlabeledDataflowData {
   periodStart: string;
   periodEnd: string;
   projects: Array<UnlabeledDataflowAlertProject>;
   unlabeledCost: number;
   labeledCost: number;
-}
-
-export class UnlabeledDataflowAlert implements Alert {
-  data: UnlabeledDataflowData;
-
-  constructor(data: UnlabeledDataflowData) {
-    this.data = data;
-  }
-
-  title = 'Add labels to workflows';
-  subtitle =
-    'Labels show in billing data, enabling cost insights for each workflow.';
-  url = '/cost-insights/labeling-jobs';
-
-  get element() {
-    return <UnlabeledDataflowAlertCard alert={this.data} />;
-  }
 }
 
 export interface UnlabeledDataflowAlertProject {
