+++ plugins/cost-insights/src/types/Currency.ts
@@ -21,3 +21,10 @@ export interface Currency {
   prefix?: string;
   rate?: number;
 }
+
+export enum CurrencyType {
+  USD = 'USD',
+  CarbonOffsetTons = 'CARBON_OFFSET_TONS',
+  Beers = 'BEERS',
+  IceCream = 'PINTS_OF_ICE_CREAM',
+}
