+++ plugins/cost-insights/src/types/index.ts
@@ -19,6 +19,7 @@ export * from './ChangeStatistic';
 export * from './ChartData';
 export * from './Cost';
 export * from './DateAggregation';
+export * from './Duration';
 export * from './Currency';
 export * from './Entity';
 export * from './Icon';
