+++ plugins/cost-insights/src/utils/duration.test.ts
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-import { Duration, inclusiveEndDateOf, inclusiveStartDateOf } from './duration';
+import { Duration } from '../types';
+import { inclusiveEndDateOf, inclusiveStartDateOf } from './duration';
 
 const lastCompleteBillingDate = '2020-06-05';
